New python script for processing bloods deltas
