--- cishouseholds/pipeline/bloods delta_ETL.py
@@ -1,16 +1,0 @@
-def bloods_delta_ETL():
-    extract_bloods_delta()
-    transform_bloods_delta()
-    load_bloods_delta()
-
-
-def extract_bloods_delta():
-    pass
-
-
-def transform_bloods_delta():
-    pass
-
-
-def load_bloods_delta():
-    pass
